requirements.txt
fastapi
uvicorn
pydantic
pathlib
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain-google-genai
sentence-transformers
langchain-community 
langchain-core
langchain-openai